+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 ## [Unreleased]
 
+## [v0.2.4] - 2019-01-14
+
+
 ### Added
 
 - `:timeout` is now supported for Strategy.Global (so it is now
@@ -57,5 +60,6 @@
 
 
 [Unreleased]: https://github.com/pragdave/component/compare/v0.2.2...HEAD
+[v0.2.4]: https://github.com/pragdave/component/compare/v0.2.3...v0.2.4
 [v0.2.3]: https://github.com/pragdave/component/compare/v0.2.2...v0.2.3
 [v0.2.2]: https://github.com/pragdave/component/compare/v0.2.1...v0.2.2